+++ Sources/Sentry/PrivateSentrySDKOnly.h
@@ -10,8 +10,12 @@ NS_ASSUME_NONNULL_BEGIN
  * ATTENTION: This class is reserved for hybrid SDKs. Methods may be changed, renamed or removed
  * without notice. If you want to use one of these methods here please open up an issue and let us
  * now.
+ *
+ * The name of this class is supposed to be a bit weird and ugly. The name starts with private on
+ * purpose so users don't see it in code completion when typing Sentry. We also add only at the end
+ * to make it more obvious you shouldn't use it.
  */
-@interface SentryHybridSDKsOnly : NSObject
+@interface PrivateSentrySDKOnly : NSObject
 
 /**
  * For storing an envelope synchronously to disk.