+++ Tests/SentryTests/PrivateSentrySDKOnlyTests.swift
@@ -1,13 +1,13 @@
 import XCTest
 
-class SentryHybridSDKsOnlyTests: XCTestCase {
+class PrivateSentrySDKOnlyTests: XCTestCase {
 
     func testStoreEnvelope() {
         let client = TestClient(options: Options())
         SentrySDK.setCurrentHub(TestHub(client: client, andScope: nil))
         
         let envelope = TestConstants.envelope
-        SentryHybridSDKsOnly.store(envelope)
+        PrivateSentrySDKOnly.store(envelope)
         
         XCTAssertEqual(1, client?.storedEnvelopes.count)
         XCTAssertEqual(envelope, client?.storedEnvelopes.first)
@@ -15,11 +15,11 @@ class SentryHybridSDKsOnlyTests: XCTestCase {
 
     func testEnvelopeWithData() throws {
         let itemData = "{}\n{\"length\":0,\"type\":\"attachment\"}\n".data(using: .utf8)!
-        XCTAssertNotNil(SentryHybridSDKsOnly.envelope(with: itemData))
+        XCTAssertNotNil(PrivateSentrySDKOnly.envelope(with: itemData))
     }
     
     func testGetDebugImages() {
-        let sut = SentryHybridSDKsOnly()
+        let sut = PrivateSentrySDKOnly()
         let images = sut.getDebugImages()
         
         // Only make sure we get some images. The actual tests are in