+++ Sources/Sentry/SentrySpanContext.m
@@ -91,6 +91,9 @@ SentrySpanContext () {
     if (self.operation != nil)
         [mutabledictionary setValue:self.operation forKey:@"op"];
 
+    if (self.spanDescription != nil)
+        [mutabledictionary setValue:self.spanDescription forKey:@"description"];
+
     if (self.parentSpanId != nil)
         [mutabledictionary setValue:self.parentSpanId.sentrySpanIdString forKey:@"parent_span_id"];