+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## Unreleased
+
+- feat: Add SentryHybridSDKsOnly (#1131)
+
 ## 7.1.2
 
 - fix: Serialization of span description (#1128)