+++ CHANGELOG.md
@@ -3,6 +3,13 @@
 ## Unreleased
 
 - feat: Add UI auto instrumentation (#1105)
+## 7.1.2
+
+- fix: Serialization of span description (#1128)
+
+## 7.1.1
+
+- No documented changes. This is the same as 7.1.0. Ignore this release and please use 7.1.2 instead.
 
 ## 7.1.0